+++ linker_hand/include/LinkerHandL20Can.h
@@ -10,7 +10,7 @@ public:
 
     void setJointPositions(const std::vector<double>& jointAngles) override;
     std::vector<double> getCurrentStatus() override;
-    void setJointSpeed(const std::vector<double>& speed) override;
+    void setJointSpeed(const std::vector<int>& speed) override;
     std::vector<double> getSpeed() override;
     std::vector<std::vector<double>> getForce() override;
     void getNormalForce() override;
